django-env/requirements.txt
Django==1.8.6
djangorestframework==3.3.1
gunicorn==19.3.0
dj-database-url==0.3.0
dj-static==0.0.6
static==0.4
wsgiref==0.1.2